# back/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "User API",
    "description": "API for managing users",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],

  "paths": {
    "/user/login": {
      "post": {
        "summary": "loggin users",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "loggin user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/signin": {
      "post": {
        "summary": "Create a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The newly created user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
    "/chat/create": {
      "post": {
        "summary": "Create a new chat",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Chat"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The newly created pix",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResp"
                }
              }
            }
          }
        }
      }
    },
    "/chat/send": {
      "post": {
        "summary": "Send a new message",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSend"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucess",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatSend"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "responses": {
      "UnauthorizedError": {
        "description": "Acesso negado token não encontrado"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "about": {
            "type": "string"
          },
          "age": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },

      "Login": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "Chat": {
        "type": "object",
        "properties": {
          "senderId": {
            "type": "integer"
          },
          "receiveId": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ChatResp": {
        "type": "object",
        "properties": {
          "conversation": {
            "type": "object"
          },
          "participants": {
            "type": "object"
          },
          "newMessage": {
            "type": "object"
          }
        }
      },
      "ChatSend": {
        "type": "object",
        "properties": {
          "conversationId": {
            "type": "integer"
          },
          "senderId": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}